requirements.txt
behave
playwright
python-dotenv
requests
pytest
PyJWT

